==> 61-WIND-TURBIN/11-turbin-basic-3-days.ipynb <==
import os
import numpy as np
import pandas as pd

from datetime import datetime

import matplotlib.pyplot as plt 
import seaborn as sns

import platform
import matplotlib.font_manager as fm

_os_name = platform.system()

# font_path는 사용할 폰트에 따라 변경하면 됨
_font_path = 'c:/Windows/Fonts/malgun.ttf' if _os_name=='Windows' \
    else '/usr/share/fonts/truetype/nanum/NanumGothic.ttf'
_font_family = fm.FontProperties(fname=_font_path).get_name()

# font 설정
plt.rcParams['font.family'] = _font_family

# 폰트캐시까지 삭제 후 한글폰트가 사용될 수 있도록 캐시 삭제
# --> 이렇게 해야 비로소 한글이 표현되는 경우가 많음
plt.rcParams['font.family'] = 'sans-serif'
plt.rcParams['font.sans-serif'] = [_font_family]

_fe = fm.FontEntry(
    fname='/usr/share/fonts/truetype/nanum/NanumGothic.ttf',
    name='NanumGothic'
)
fm.fontManager.ttflist.insert(0, _fe)

_PATH_BASE = os.path.join(os.getcwd(), 'data', '41-turbin')
get_path = lambda turbin: os.path.join(_PATH_BASE, f'{turbin}_turbin.csv')

df_data = pd.read_csv(get_path('b'))

# 처음 2개와 마지막 4개
df_data = df_data.iloc[2:-4]

# 기존 인덱스를 컬럼으로 변경하고 새 인덱스를 생성함
df_data.reset_index(inplace=True)
# 기존 인덱스 제거
df_data.drop('index', axis=1, inplace=True)

df_data_back = df_data.copy()

target_col = 'ACTIVE_POWER'

# 결측치 처리
# 발전량이 음수인 컬럼의 풍향/풍속/발전량 값을 NaN으로 변경
df_data.loc[df_data[target_col] < 0, ['WIND_SPEED', 'WIND_DIR']+[target_col]] = np.nan

# 시계열성값을 이용한 결측치 처리
df_data = df_data.interpolate()

df_data.isna().sum()

df_temp = df_data.head(24*10)

df_temp.info()

df_temp['TURBINE_TIME'] = pd.to_datetime(df_temp['TURBINE_TIME'])
df_temp.set_index('TURBINE_TIME', inplace=True)
df_temp.index.name='index'

df_temp.info()

df_temp_daily = df_temp.resample('D').agg({
    'WIND_SPEED': ['mean', 'sum', 'min', 'max'], 
    'WIND_DIR': ['mean', 'sum', 'min', 'max'], 
    'ACTIVE_POWER': ['mean', 'sum', 'min', 'max']
})

df_temp_daily.columns = [f'{col}_{stat}' for col, stat in df_temp_daily.columns]

df_temp_daily.head()
