==> 61-WIND-TURBIN/21-turbin-pp.ipynb <==
#### 라이브러리 불러오기

import os

import numpy as np
import pandas as pd

_PATH_BASE = os.path.join(os.getcwd(), 'data')
get_path = lambda turbin: os.path.join(_PATH_BASE, f'{turbin}_turbin.csv')
df_data = pd.read_csv(get_path('b'))

#### 하루단위가 안되는 ROW 삭제

df_data = df_data.iloc[2:-4]

# 기존 인덱스 갱신
df_data.reset_index(inplace=True)
df_data.drop('index', axis=1, inplace=True)

target_col = 'ACTIVE_POWER'
df_data[df_data[target_col]==-1].shape

non_value_counts = (df_data.drop('TURBINE_TIME', axis=1) < 0).sum()
non_value_counts

# 발전량이 음수인 컬럼의 풍향/풍속/발전량 값을 NaN으로 변경
df_data.loc[df_data[target_col] < 0, ['WIND_SPEED', 'WIND_DIR']+[target_col]] = np.nan

# 시계열성값을 이용한 결측치 처리
df_data = df_data.interpolate()

#### 풍향/풍속을 이용해 바람벡터 생성

ws = df_data.pop('WIND_SPEED')
wd = df_data.pop('WIND_DIR')
wd_radian = wd * np.pi /180

df_data['ws_X'] = ws * np.cos(wd_radian)
df_data['ws_y'] = ws * np.sin(wd_radian)

#### 시간데이터에 일간/년간 주기성 부여

day = 24 * 60 * 60
year = 365.2425 * day
df_data['TURBINE_TIME'] = pd.to_datetime(df_data['TURBINE_TIME'])
df_data.set_index('TURBINE_TIME', inplace=True)

timestampe_s = df_data.index.map(pd.Timestamp.timestamp)
df_data['day_sin'] = np.sin(timestampe_s * (2 * np.pi / day))
df_data['day_cos'] = np.cos(timestampe_s * (2 * np.pi / day))
df_data['year_sin'] = np.sin(timestampe_s * (2 * np.pi / year))
df_data['year_cos'] = np.cos(timestampe_s * (2 * np.pi / year))

df_data.drop(columns=['month', 'hour'], axis=1, inplace=True)

data_size, n_features = df_data.shape

df_train = df_data[:int(data_size * 0.7)]
df_valid = df_data[int(data_size * 0.7): int(data_size * 0.9)]
df_test = df_data[int(data_size * 0.9):] 

data_mean, data_std = df_train.mean(), df_train.std()

df_train = (df_train - data_mean) / data_std
df_valid = (df_valid - data_mean) / data_std
df_test = (df_test - data_mean) / data_std

df_train.to_csv(os.path.join(_PATH_BASE, 'b_train.csv'), index=False)
df_valid.to_csv(os.path.join(_PATH_BASE, 'b_valid.csv'), index=False)
df_test.to_csv(os.path.join(_PATH_BASE, 'b_test.csv'), index=False)
