==> 61-WIND-TURBIN/32-turbin-daily-win.ipynb <==
import os
import numpy as np
import pandas as pd
import tensorflow as tf

_PATH_BASE = os.path.join(os.getcwd(), 'data')
df_daily = pd.read_csv(
    os.path.join(_PATH_BASE, 'b_daily.csv'), 
    index_col=None
)

df_daily.drop('index', axis=1, inplace=True)

WINDOW_SIZE = 7
target_col = 'ACTIVE_POWER_sum'
column_indices = {name: i for i, name in enumerate(df_daily.columns)}
target_idx = column_indices[target_col]

# Features, Target 분류
X_data_win = []
y_data_win = []
nd_daily = df_daily.values

for _i in range(df_daily.shape[0] - WINDOW_SIZE):
    X_data_win.append(nd_daily[_i:_i+WINDOW_SIZE, :])
    y_data_win.append(nd_daily[_i+WINDOW_SIZE, target_idx])

X_data_win = np.array(X_data_win)
y_data_win = np.array(y_data_win)

np.save(os.path.join(_PATH_BASE, 'b_daily_feature.npy'), X_data_win)
np.save(os.path.join(_PATH_BASE, 'b_daily_label.npy'), y_data_win)

def make_window(data, window_size, batch_size, shuffle):
    ds = tf.data.Dataset.from_tensor_slices(data)
    ds = ds.window(window_size+1, shift=1, drop_remainder=True)
    

ds = tf.data.Dataset.from_tensor_slices(nd_daily)

ds = ds.window(size=WINDOW_SIZE, shift=1, stride=1, drop_remainder=True)

ds

132*7

temp_ndarray = np.array([[1,2,3], [4,5,6], [7,8,9]])

def test_func(x):
    print(x)
    tf.print(x)
    return tf.data.Dataset.from_tensor_slices(x)

temp_ds = test_func(temp_ndarray)
list(temp_ds.as_numpy_iterator())

temp_ds2 = temp_ds.flat_map(test_func)
list(temp_ds2.as_numpy_iterator())

tf.print(temp_ds)
