==> 61-WIND-TURBIN/52-ts-climate-pp.ipynb <==
#### 라이브러리 불러오기

import os

import numpy as np
import pandas as pd

#### 데이터 불러오기

_PATH_BASE = os.path.join(os.getcwd(), 'data')
_PATH_DATA_FILE = os.path.join(_PATH_BASE, 'climate_2009_2016.csv')

if os.path.exists(_PATH_DATA_FILE):
    df_climate = pd.read_csv(_PATH_DATA_FILE)
else:
    _WEB_FILE_PATH = tf.keras.utils.get_file(
        origin='https://storage.googleapis.com/tensorflow/tf-keras-datasets/jena_climate_2009_2016.csv.zip',
        fname='jena_climate_2009_2016.csv.zip',
        extract=True
    )
    _CSV_PATH, _ = os.path.splitext(_WEB_FILE_PATH)
    df_climate = pd.read_csv(_CSV_PATH)
    df_climate.to_csv(_PATH_DATA_FILE, index=False)

#### 데이터 전처리

##### 10분단위 데이터 -> 시간단위 데이터

# 10분단위로 들어온 데이터를 시간당 한건만 추출하는 방법
# df_climate = df_climate[5::6]     # 5번째 로우부터 6로우(60분) 간격으로 하나씩 추출

# 해당시간대의 평균값으로 시간별 데이터 추출
df_climate['Date Time'] = \
    pd.to_datetime(df_climate['Date Time'], format='%d.%m.%Y %H:%M:%S')
    
# 시계열데이터가 인덱스가 아니면, 아래 resample()이 처리되지 않음
df_climate.set_index('Date Time', inplace=True)
df_climate = df_climate.resample('H').mean()

##### 풍속/최대풍속 음수값 처리

# 아래 두 행은 같은 방법임
# 이 두 행은 조건에 의해 특정 컬럼값을 변경시키는 방법임(같은 기능)
df_climate['wv (m/s)'][df_climate['wv (m/s)'] < 0.0] = 0.0
# df_climate.loc[df_climate['wv (m/s)'] < 0.0, 'wv (m/s)'] = 0.0
df_climate['max. wv (m/s)'] = df_climate['max. wv (m/s)'].apply(lambda x: max(0,x))

df_climate.describe().T

##### 풍향/풍속을 이용해 바람벡터 생성

df_climate.shape

_wv = df_climate.pop('wv (m/s)')
_max_wv = df_climate.pop('max. wv (m/s)')

_wd_radian = df_climate.pop('wd (deg)') * np.pi / 180

# 바람백터 생성
df_climate['Vwv_X'] = _wv * np.cos(_wd_radian)
df_climate['Vwv_Y'] = _wv * np.sin(_wd_radian)
df_climate['Vmwv_X'] = _max_wv * np.cos(_wd_radian)
df_climate['Vmwv_Y'] = _max_wv * np.sin(_wd_radian)

# 3개 컬럼이 제거되고 4개 컬럼이 추가됨
df_climate.shape

df_climate.info()

##### 시간데이터에 일간/년간 주기성 부여

# 시간데이터가 초 단위로 있으면 모델링에 적합하지 않다.
# 일간/년간 주기성 부여가 반드시 필요함
_day = 24 * 60 * 60
_year = (365.2425) * _day
_timestamp_s = df_climate.index.map(pd.Timestamp.timestamp)

df_climate['day_sin'] = np.sin(_timestamp_s * (2 * np.pi / _day))
df_climate['day_cos'] = np.cos(_timestamp_s * (2 * np.pi / _day))
df_climate['year_sin'] = np.sin(_timestamp_s * (2 * np.pi / _year))
df_climate['year_cos'] = np.cos(_timestamp_s * (2 * np.pi / _year))

##### 데이터 분할

_data_size, _n_features = df_climate.shape

df_train = df_climate[:int(_data_size * 0.7)]
df_valid = df_climate[int(_data_size * 0.7): int(_data_size * 0.9)]
df_test = df_climate[int(_data_size * 0.9):]

##### 정규화

_data_mean, _data_std = df_train.mean(), df_train.std()

df_train = (df_train - _data_mean) / _data_std
df_valid = (df_valid - _data_mean) / _data_std
df_test = (df_test - _data_mean) / _data_std

##### 전처리 데이터 저장

df_train.to_csv(os.path.join(_PATH_BASE, 'climate_train.csv'), index=False)
df_valid.to_csv(os.path.join(_PATH_BASE, 'climate_valid.csv'), index=False)
df_test.to_csv(os.path.join(_PATH_BASE, 'climate_test.csv'), index=False)
df_climate.to_csv(os.path.join(_PATH_BASE, 'climate_pp.csv'), index=False)
