==> 61-WIND-TURBIN/54-ts-climate-win-test.ipynb <==
import os
import pandas as pd

from climate_win_gen import WindowGen

w1 = WindowGen(input_width=1, label_width=1, shift=1)

_PATH_BASE = os.path.join(os.getcwd(), 'data')
_PATH_TRAIN = os.path.join(_PATH_BASE, 'climate_train.csv')
_PATH_VALID = os.path.join(_PATH_BASE, 'climate_valid.csv')
_PATH_TEST = os.path.join(_PATH_BASE, 'climate_test.csv')

try:
    df_train = pd.read_csv(_PATH_TRAIN)
    df_valid = pd.read_csv(_PATH_VALID)
    df_test = pd.read_csv(_PATH_TEST)
except Exception as e:
    print('훈련/검증/시험파일이 존재하지 않습니다.\n'
          '52-ts-climate-pp.ipynb 실행 후 작업하시기 바랍니다.')
    print(e)

def make_win(input_width, label_width, shift, labels):
    return WindowGen(
        input_width=input_width,
        label_width=label_width,
        shift=shift,
        df_train=df_train,
        df_valid=df_valid,
        df_test=df_test,
        label_columns=labels
    )

w1 = make_win(input_width=1, label_width=1, shift=1, labels=['T (degC)'])
w1
