==> 61-WIND-TURBIN/56-ts-climate-modeling.ipynb <==
 import os
 import numpy as np
 import pandas as pd
 import tensorflow as tf
 
 import matplotlib.pyplot as plt
 
 from freeman.plt_setting import plt_settings
 from freeman.tensor import MakeWindow

plt_settings()

_PATH_BASE = os.path.join(os.getcwd(), 'data')
_PATH_CLIMATE = os.path.join(_PATH_BASE, 'climate_pp.csv')

try:
    df_climate = pd.read_csv(_PATH_CLIMATE)
except Exception as e:
    print('파일이 존재하지 않습니다.\n'
          '52-ts-climate-pp.ipynb 실행 후 작업하시기 바랍니다.')
    print(e)

WINDOW_SIZE = 24
BATCH_SIZE = 32
TARGET_COL = 'T (degC)'

win_data = MakeWindow(
    df_climate, WINDOW_SIZE, BATCH_SIZE, label_col=TARGET_COL
)

win_data.slices()

model = tf.keras.Sequential([
    tf.keras.layers.LSTM(32, activation='tanh', input_shape=win_data.input_shape),
    tf.keras.layers.Dense(1)
])

model.compile(
    optimizer=tf.keras.optimizers.Adam(),
    loss=tf.keras.losses.MeanSquaredError(),
    metrics=[tf.keras.metrics.MeanAbsoluteError()]
)

model.fit(       # return history
    win_data.train, 
    epochs=10, 
    validation_data=win_data.valid,
)

model.evaluate(win_data.test)
